Winrate fix.

+++ database/migrations/2024_02_04_155241_create_leaderboard_onevones_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->integer('wins');
             $table->integer('losses');
             $table->integer('ties');
-            $table->integer('winrate');
+            $table->float('winrate');
             $table->timestamps();
         });
